Move NodeJS tests out of `pkg/codegen`
This commit moves the NodeJS tests into
`pkg/codegen/testing/test/nodejs.go`, so that they won't be included in
`pkg/codegen` code by downstream dependencies.
