+++ pkg/codegen/nodejs/gen_test.go
@@ -62,7 +62,7 @@ func testGeneratePackageBatch(t *testing.T, testCases []*test.SDKTest) {
 		},
 		Checks: map[string]test.CodegenCheck{
 			"nodejs/compile": func(t *testing.T, pwd string) {
-				typeCheckGeneratedPackage(t, pwd, true)
+				test.TypeCheckNodeJSPackage(t, pwd, true)
 			},
 			"nodejs/test": testGeneratedPackage,
 		},
